integrate scenes
